+++ apps/frontend/src/libs/enums/enums.ts
@@ -2,10 +2,21 @@ export { AppRoute } from "./app-route.enum.js";
 export { AudioEvent } from "./audio-event.enum.js";
 export { AvatarSize } from "./avatar-size.enum.js";
 export { AvatarType } from "./avatar-type.enum.js";
+export { Breakpoint } from "./breakpoint.enum.js";
 export { ButtonSize } from "./button-size.enum.js";
 export { ButtonVariant } from "./button-variant.enum.js";
+export { CarouselConfig } from "./carousel-config.enum.js";
+export { ContactInfo } from "./contact-info.enum.js";
 export { DataStatus } from "./data-status.enum.js";
+export { DOMEvent } from "./dom-event.enum.js";
+export { FeatureAnimation } from "./feature-animation.enum.js";
 export { InputPasswordType } from "./input-password.enum.js";
+export {
+	LandingBgRingType,
+	LandingBgSide,
+} from "./landing-background.enums.js";
+export { LandingSection } from "./landing-section.enum.js";
+export { LandingTestimonialConfig } from "./landing-testimonial.enum.js";
 export { LoaderIconConfig } from "./loader-icon-config.enum.js";
 export { LoaderProgressConfig } from "./loader-progress-config.enum.js";
 export { LogoSize } from "./logo-size.enum.js";