+++ apps/frontend/src/libs/types/in-view-animation-options.type.ts
@@ -1,0 +1,11 @@
+type InViewAnimationOptions = {
+	amount?: number;
+	duration?: number;
+	final: Partial<CSSStyleDeclaration>;
+	initial: Partial<CSSStyleDeclaration>;
+	isDisabled?: boolean;
+	margin?: string;
+	ref: React.RefObject<HTMLElement>;
+};
+
+export { type InViewAnimationOptions };