+++ packages/shared/src/libs/types/date-format.type.ts
@@ -1,3 +1,3 @@
-type DateFormat = "D MMMM hA" | "mm:ss" | "MMM D, YYYY, h:m A";
+type DateFormat = "D MMMM hA" | "mm:ss" | "MMM D, YYYY, h:mm A";
 
 export { type DateFormat };